Unify db classes between MS and requrest collector
