--- src/request_collector/state_db/__init__.py
@@ -1,5 +1,0 @@
-from .sqlite_db import StateDBSqlite
-
-__all__ = [
-    'StateDBSqlite',
-]
