+++ tests/monitoring/fixtures/__init__.py
@@ -1,4 +1,3 @@
 from .contracts import *    # noqa
 from .monitor_request import *  # noqa
-from .state_db import *  # noqa
 from .server import *  # noqa
