+++ tests/monitoring/fixtures/monitor_request.py
@@ -11,14 +11,14 @@ def get_monitor_request_for_same_channel(
         get_random_address,
         get_random_privkey,
         token_network,
-        state_db_sqlite,
+        ms_database,
 ):
     keys = [get_random_privkey() for i in range(3)]
     token_network_address = token_network.address
 
     channel_id = 1
     balance_hash_data = '0'
-    state_db_sqlite.upsert_channel(Channel(
+    ms_database.upsert_channel(Channel(
         identifier=channel_id,
         token_network_address=token_network_address,
         participant1=private_key_to_address(keys[0]),
